requirements.txt
Django==3.1.6
djangorestframework==3.12.2
pytz==2021.1
asgiref==3.3.1
python-decouple==3.4
sqlparse==0.4.1